Fix playTrack
