+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/player/music/MusicUnlock.kt
@@ -1,10 +1,7 @@
 package world.gregs.voidps.world.interact.entity.player.music
 
-import world.gregs.voidps.engine.client.playMusicTrack
-import world.gregs.voidps.engine.client.ui.sendText
+import world.gregs.voidps.engine.client.ui.playTrack
 import world.gregs.voidps.engine.client.variable.addVarbit
-import world.gregs.voidps.engine.client.variable.set
-import world.gregs.voidps.engine.data.definition.extra.EnumDefinitions
 import world.gregs.voidps.engine.data.definition.extra.VariableDefinitions
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.get
@@ -21,11 +18,3 @@ object MusicUnlock {
 fun Player.unlockTrack(track: String) = addVarbit("unlocked_music_${get<MusicTracks>().get(track) / 32}", track)
 
 fun Player.playTrack(trackName: String) = playTrack(get<MusicTracks>().get(trackName))
-
-fun Player.playTrack(trackIndex: Int) {
-    val enums: EnumDefinitions = get()
-    playMusicTrack(enums.get("music_tracks").getInt(trackIndex))
-    val name = enums.get("music_track_names").getString(trackIndex)
-    interfaces.sendText("music_player", "currently_playing", name)
-    this["current_track"] = trackIndex
-}
